backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
resend>=2.5.1
sentry-sdk[fastapi]>=2.19.0
python-multipart>=0.0.20
psycopg2-binary>=2.9.10
aiosqlite>=0.20.0
pydantic[email]>=2.0.0
